requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.9.24
cfgv==3.3.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
distlib==0.3.6
Django==4.1.3
django-health-check==3.17.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
filelock==3.8.0
identify==2.5.8
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
nodeenv==1.7.0
packaging==21.3
Pillow==9.3.0
platformdirs==2.5.3
pre-commit==2.20.0
psutil==5.9.4
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.6
PyYAML==6.0
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
sqlparse==0.4.3
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.6
wrapt==1.14.1
